fix: getproductdetails to features
