--- finpro-web/src/services/getProductDetails.ts
@@ -1,12 +1,0 @@
-import { IProductDetails } from "@/types/products/product.type";
-import apiInstance from "./apiInstance";
-
-export const getProductDetails = async (params: string) => {
-    try {
-        const response = await apiInstance.get("/product/" + params);
-        return response.data.product as IProductDetails;
-    } catch (error) {
-        console.error("Error fetching product details:", error);
-        throw error;
-    }
-}
